feat: add payment notification endpoint to PaymentController
