+++ src/main/java/com/DevSprint/voluntrix_backend/controllers/PaymentController.java
@@ -1,13 +1,17 @@
 package com.DevSprint.voluntrix_backend.controllers;
 
+import com.DevSprint.voluntrix_backend.dtos.PayHereNotificationDTO;
 import com.DevSprint.voluntrix_backend.dtos.PaymentRequestDTO;
 import com.DevSprint.voluntrix_backend.dtos.PaymentResponseDTO;
 import com.DevSprint.voluntrix_backend.dtos.PaymentStatusResponseDTO;
+import com.DevSprint.voluntrix_backend.exceptions.PaymentVerificationException;
 import com.DevSprint.voluntrix_backend.services.PaymentService;
+import com.DevSprint.voluntrix_backend.utils.PaymentMapper;
 
-import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
+
+import java.util.Map;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -15,17 +19,18 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.bind.annotation.RequestBody;
 
 @Validated
 @RequiredArgsConstructor
 @RestController
-@SecurityRequirement(name = "bearerAuth")
 @RequestMapping("/api/payment")
 public class PaymentController {
 
     private final PaymentService paymentService;
+    private final PaymentMapper paymentMapper;
 
     @PostMapping("/start")
     public ResponseEntity<PaymentResponseDTO> startPayment(@RequestBody @Valid PaymentRequestDTO paymentRequest) {
@@ -37,4 +42,18 @@ public class PaymentController {
         PaymentStatusResponseDTO statusDto = paymentService.getPaymentStatusByOrderId(orderId);
         return ResponseEntity.ok(statusDto);
     }
+
+    @PostMapping("/payment/notify")
+    public ResponseEntity<String> notifyPayment(@RequestParam Map<String, String> params) {
+        boolean isValid = paymentService.verifyPayment(params);
+
+        if (!isValid) {
+            throw new PaymentVerificationException("Payment signature verification failed");
+        }
+
+        PayHereNotificationDTO dto = paymentMapper.toNotificationDto(params);
+        paymentService.saveTransaction(dto);
+        
+        return ResponseEntity.ok("Transaction saved succesfully.");
+    } 
 }
