+++ src/main/java/com/DevSprint/voluntrix_backend/controllers/PublicController.java
@@ -4,27 +4,21 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.DevSprint.voluntrix_backend.dtos.EventDTO;
 import com.DevSprint.voluntrix_backend.dtos.InstituteDTO;
-import com.DevSprint.voluntrix_backend.dtos.PayHereNotificationDTO;
 import com.DevSprint.voluntrix_backend.dtos.SocialFeedResponseDTO;
 import com.DevSprint.voluntrix_backend.dtos.SponsorDTO;
 import com.DevSprint.voluntrix_backend.dtos.VolunteerDTO;
 import com.DevSprint.voluntrix_backend.enums.UserType;
-import com.DevSprint.voluntrix_backend.exceptions.PaymentVerificationException;
 import com.DevSprint.voluntrix_backend.services.EventService;
 import com.DevSprint.voluntrix_backend.services.InstituteService;
-import com.DevSprint.voluntrix_backend.services.PaymentService;
 import com.DevSprint.voluntrix_backend.services.SocialFeedService;
 import com.DevSprint.voluntrix_backend.services.SponsorService;
 import com.DevSprint.voluntrix_backend.services.VolunteerService;
 import com.DevSprint.voluntrix_backend.utils.ApiResponse;
-import com.DevSprint.voluntrix_backend.utils.PaymentMapper;
 import com.DevSprint.voluntrix_backend.validation.RequiresRole;
 
 import io.swagger.v3.oas.annotations.Parameter;
@@ -33,7 +27,6 @@ import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 
 import java.util.List;
-import java.util.Map;
 
 @RequiredArgsConstructor
 @RestController
@@ -47,8 +40,6 @@ public class PublicController {
     private final VolunteerService volunteerService;
     private final EventService eventService;
     private final SocialFeedService socialFeedService;
-    private final PaymentService paymentService;
-    private final PaymentMapper paymentMapper;
 
     @GetMapping("/institutes")
     @RequiresRole({UserType.VOLUNTEER, UserType.SPONSOR, UserType.ORGANIZATION, UserType.ADMIN, UserType.PUBLIC})
@@ -91,18 +82,4 @@ public class PublicController {
         List<SocialFeedResponseDTO> posts = socialFeedService.getAllPosts();
         return ResponseEntity.ok(posts);
     }
-
-    @PostMapping("/payment/notify")
-    public ResponseEntity<String> notifyPayment(@RequestParam Map<String, String> params) {
-        boolean isValid = paymentService.verifyPayment(params);
-
-        if (!isValid) {
-            throw new PaymentVerificationException("Payment signature verification failed");
-        }
-
-        PayHereNotificationDTO dto = paymentMapper.toNotificationDto(params);
-        paymentService.saveTransaction(dto);
-        
-        return ResponseEntity.ok("Transaction saved succesfully.");
-    } 
 }
